Move code to ee folder
