+++ app/api/server/index.js
@@ -37,7 +37,6 @@ import './v1/webdav';
 import './v1/oauthapps';
 import './v1/custom-sounds';
 import './v1/custom-user-status';
-import './v1/licenses';
 import './v1/instances';
 
 export { API, APIClass, defaultRateLimiterOptions } from './api';
