+++ ee/server/api/index.ts
@@ -1,0 +1,1 @@
+import './licenses';
