+++ ee/server/index.js
@@ -10,3 +10,4 @@ import '../app/engagement-dashboard/server/index';
 import '../app/ldap-enterprise/server/index';
 import '../app/livechat-enterprise/server/index';
 import '../app/settings/server/index';
+import './api';
